requirements.txt
Flask
Flask-Cors
firebase-admin
qrcode
pyOpenSSL
cryptography
requests
pyserial
